[feat] complete 속성 Volume 모델로 이동 및 getOneRM query 재디자인
#131
