+++ src/models/Volume.ts
@@ -6,15 +6,22 @@ import { CardiovascularVolume } from './CardiovascularVolume';
 import { Model } from './Model';
 import { Plan } from './Plan';
 import { WeightVolume } from './WeightVolume';
-import { setTotal } from './hooks/volume-hooks';
+import { setOneRMWithTotal } from './hooks/volume-hooks';
 import { VolumeMethods, VolumeQueryHelpers } from './types/Volume';
 
-@pre<Volume>('save', setTotal)
+@pre<Volume>('save', setOneRMWithTotal)
 @ObjectType({ implements: Model, description: '운동 볼륨' })
 export class Volume extends Model implements VolumeMethods {
   @Field(() => Plan, { description: '운동계획' })
   @prop({ ref: 'Plan', required: true })
   plan: Ref<Plan>;
+
+  @Field(() => Boolean, {
+    description: '완료 여부',
+    defaultValue: false,
+  })
+  @prop({ type: Boolean, default: false })
+  complete: boolean;
 
   @Field(() => Int, { description: '횟수', nullable: true })
   @prop({ type: Number })
@@ -34,6 +41,10 @@ export class Volume extends Model implements VolumeMethods {
 
   @Field(() => Float, { description: '총 볼륨', defaultValue: 0 })
   total: number;
+
+  @Field(() => Float, { description: '1rm', defaultValue: 0 })
+  @prop({ type: Number, default: 0 })
+  oneRM: number;
 
   isCalisthenicsVolume(): this is CalisthenicsVolume {
     return (
