+++ src/models/hooks/volume-hooks.ts
@@ -1,6 +1,11 @@
-import { Volume } from '@src/models/Volume';
 import { PreFnWithDocumentType } from '@src/types/hooks';
 
-export const setTotal: PreFnWithDocumentType<Volume> = async function () {
-  this.total = this.isWeightVolume() ? this.weight * this.count : 0;
-};
+import { Volume } from '../Volume';
+
+export const setOneRMWithTotal: PreFnWithDocumentType<Volume> =
+  async function () {
+    this.oneRM = this.isWeightVolume()
+      ? this.weight + this.weight * this.count * 0.025
+      : 0;
+    this.total = this.isWeightVolume() ? this.weight * this.count : 0;
+  };
