+++ src/resolvers/PlanResolver.ts
@@ -40,33 +40,6 @@ export class PlanResolver implements ResolverInterface<Plan> {
     return PlanModel.find({ user: user._id });
   }
 
-  @Query(() => Number, { description: '최대 무게' })
-  @UseMiddleware(AuthenticateMiddleware)
-  async getOneRM(
-    @Arg('name') name: string,
-    @Ctx() { user }: Context,
-  ): Promise<number> {
-    if (!user) {
-      throw new AuthenticationError();
-    }
-
-    const training = await TrainingModel.findOne({ name });
-
-    if (training === null) {
-      return 0;
-    }
-
-    return (
-      (
-        await PlanModel.findOne({
-          user: user._id,
-          training: training._id.toHexString(),
-          complete: true,
-        }).sort({ oneRM: -1 })
-      )?.oneRM || 0
-    );
-  }
-
   @Mutation(() => [Plan], { description: '여러개의 운동 계획 생성 및 수정' })
   @UseMiddleware(AuthenticateMiddleware)
   async multipleCreateOrUpdatePlans(
