+++ tests/unit/plan.test.ts
@@ -1,4 +1,3 @@
-import DocumentNotFoundError from '@src/errors/DocumentNotFoundError';
 import { PlanFactory } from '@src/factories/PlanFactory';
 import { VolumeFactory } from '@src/factories/VolumeFactory';
 import { Model } from '@src/models/Model';
@@ -26,20 +25,5 @@ describe('운동계획 모델', () => {
     await PlanModel.findByIdAndDelete(plan._id);
 
     expect(await VolumeModel.count()).toEqual(0);
-  });
-
-  it('운동 계획을 생성할 때 1rm을 저장하는 훅이 있다', async () => {
-    const { user } = await signIn();
-    const { _id } = await PlanModel.createWithVolumes(
-      user,
-      await PlanFactory({
-        volumes: [VolumeFactory({ weight: 100, count: 5 })],
-      }),
-    );
-
-    expect(
-      (await PlanModel.findById(_id).orFail(new DocumentNotFoundError()).exec())
-        .oneRM !== 0,
-    ).toBeTruthy();
   });
 });
