+++ tests/unit/volume.test.ts
@@ -11,21 +11,20 @@ describe('운동볼륨 모델', () => {
     expect(Object.getPrototypeOf(Volume)).toEqual(Model);
   });
 
-  it('중량 볼륨을 생성할 때 총볼륨을 저장하는 훅이 있다', async () => {
+  it('중량 볼륨을 생성할 때 총볼륨과 1rm을 저장하는 훅이 있다', async () => {
     const { user } = await signIn();
     const plan = await PlanModel.createWithVolumes(
       user,
       await PlanFactory({
-        volumes: [VolumeFactory({ weight: 100, count: 5 })],
+        volumes: [VolumeFactory({ weight: 100, count: 5, complete: true })],
       }),
     );
 
-    expect(
-      (
-        await VolumeModel.findById(plan.volumes[0]?._id)
-          .orFail(new DocumentNotFoundError())
-          .exec()
-      ).total !== 0,
-    ).toBeTruthy();
+    const volume = await VolumeModel.findById(plan.volumes[0]?._id)
+      .orFail(new DocumentNotFoundError())
+      .exec();
+
+    expect(volume.total !== 0).toBeTruthy();
+    expect(volume.oneRM !== 0).toBeTruthy();
   });
 });
